Moved middleware package to pkg dir
